Add "lazy" everythingWithContextBut
The original `every...` was eager because it evaluated the mapper first,
then the recursion, then the fold. This latest addition evaluates the
recursion first, then the mapper, then the fold. The difference is that
the mapper sees the updated value from the recursion, rather than v.v.
While the original is more compact, the newer one follows the same
behavior as `everythingBut`.